new package: cmd/pprof
Updates LP #1519473

This PR adds a new package, github.com/juju/juju/cmd/pprof, a fork of
net/http/pprof. The fork is necessary because importing the latter package
automatically registers debug handlers with the default server mux, which
juju does not go to lengths to protect, and exposing this data over http
will set of an explosion of security hand wringing.

Instead, the package provides a way to export the same handlers over a
unix socket owned by the user starting the process. To access the debug
statistics, which are read only, you must have physical access to the host.

These statistics are only enabled in cmd/jujud when it is acting as jujud,
a machine or unit agent.